[Refactor] Refactor the entire codebase
