+++ analysis/__init__.py
@@ -1,0 +1,4 @@
+from .video_analyzer import VideoAnalyzer
+
+__all__ = ["VideoAnalyzer"]
+
