+++ grading/__init__.py
@@ -1,0 +1,4 @@
+from .player_grader import PlayerGrader
+
+__all__ = ["PlayerGrader"]
+
