+++ pose/__init__.py
@@ -1,0 +1,4 @@
+from .detector import PoseDetector
+
+__all__ = ["PoseDetector"]
+
