+++ processing/__init__.py
@@ -1,0 +1,4 @@
+from .video_processor import VideoProcessor
+
+__all__ = ["VideoProcessor"]
+
