+++ tests/test_grader.py
@@ -1,15 +1,15 @@
 import pytest
 from unittest.mock import patch, MagicMock
-from Grader import Grader
-from ServeGrader import ServeGrader, serve_angle_grader
-from Types import AngleDict
-from GraderRegistry import GraderRegistry
-from Types import Skill, Handedness, GraderResult
+from graders.base import Grader
+from graders.serve import ServeGrader, serve_angle_grader
+from core.types import AngleDict
+from graders.registry import GraderRegistry
+from core.types import Skill, Handedness, GraderResult
 
 
 class TestServeAngleGrader:
-    @patch("ServeGrader.serve_mean")
-    @patch("ServeGrader.serve_std")
+    @patch("graders.serve.serve_mean")
+    @patch("graders.serve.serve_std")
     def test_serve_angle_grader_within_range(self, mock_std, mock_mean):
         mock_mean_loc = MagicMock()
         mock_mean_loc.__getitem__.return_value = 80  # Expected mean value
@@ -24,8 +24,8 @@ class TestServeAngleGrader:
 
         assert result == 10
 
-    @patch("ServeGrader.serve_mean")
-    @patch("ServeGrader.serve_std")
+    @patch("graders.serve.serve_mean")
+    @patch("graders.serve.serve_std")
     def test_serve_angle_grader_below_range(self, mock_std, mock_mean):
         mock_mean_loc = MagicMock()
         mock_mean_loc.__getitem__.return_value = 90  # Expected mean value
@@ -86,7 +86,7 @@ class TestServeRightHandedGrader:
         result = self.grader.grade_checkpoint_3(angle_dict)
         assert result == 20
 
-    @patch("ServeGrader.serve_angle_grader")
+    @patch("graders.serve.serve_angle_grader")
     def test_grade_checkpoint_4_calls_serve_angle_grader(self, mock_grader):
         mock_grader.return_value = 15
         angle_dict: AngleDict = {"Right Elbow": 120}
@@ -96,7 +96,7 @@ class TestServeRightHandedGrader:
         mock_grader.assert_called_once_with(20, "Right Elbow", "check4", angle_dict)
         assert result == 15
 
-    @patch("ServeGrader.serve_angle_grader")
+    @patch("graders.serve.serve_angle_grader")
     def test_grade_returns_grading_outcome(self, mock_grader):
         mock_grader.return_value = 10
