+++ video/constants.py
@@ -1,0 +1,6 @@
+SMOOTHING_WINDOW_SIZE = 5
+PEAK_ACCELERATION_OFFSET = 2
+IMPACT_FRAME_SEARCH_WINDOW_BEFORE = 30
+IMPACT_FRAME_SEARCH_WINDOW_AFTER = 30
+ANALYSIS_WINDOW_PADDING_BEFORE = 30
+
